refactor: mock 데이터 및 핸들러 폴더로 분리
