+++ front/src/mocks/browser.ts
@@ -1,6 +1,5 @@
 import { setupWorker } from 'msw';
-
-import handlers from './handlers';
+import handlers from './handlers/handlers';
 
 const worker = setupWorker(...handlers);
